+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Added
 - New interactive visualizations (#765).
 - Added option to reverse the qubit order when plotting a circuit. (#762, #786)
 - Jupyter notebook magic function qiskit_job_status (#734).
+- Add a new function ``qobj_to_circuits`` to convert a Qobj object to
+  a list of QuantumCircuit objects (#877)
 
 Changed
 -------